move remaining types code to deprecated namespace
