+++ db/deprecated/types/exceptions.py
